Use extracted lib
